+++ packages/components/src/index.ts
@@ -5,3 +5,4 @@ export * from './components/toggle/index.js';
 export * from './components/input/index.js';
 export * from './components/select/index.js';
 export * from './components/loader/index.js';
+export * from './components/tabs/index.js';